move logger to class file
