+++ DET/DETAlgs/logger.py
@@ -1,0 +1,7 @@
+class Logger:
+    def __init__(self, verbose=True):
+        self.verbose = verbose
+
+    def log(self, message):
+        if self.verbose:
+            print(message)
